backend/config/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    SECRET_KEY = os.getenv("SECRET_KEY", "default-secret-key")
    UV_API_KEY = os.getenv("UV_API_KEY", "")
    DEBUG = os.getenv("FLASK_ENV") == "development"